Moved model free designs to core.initial_designs package
