+++ emukit/core/initial_designs/__init__.py
@@ -1,5 +1,5 @@
 # Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-
-from .experimental_design_loop import ExperimentalDesignLoop  # noqa: F401
+from .latin_design import LatinDesign  # noqa: F401
+from .random_design import RandomDesign  # noqa: F401
