+++ emukit/core/initial_designs/latin_design.py
