+++ emukit/core/initial_designs/random_design.py
